// src/app/models/Sales.ts
import { SalesItems } from "./SalesItems";
export class Sales {
    clientName: string;
    cashSalesNo: string;
    address: string;
    contactNo: string;
    email: string;
    date: string;
    duedate: string;
    currency: string;
    paymentType: String;
    subTotal: string;
    paidAmount: string;
    changes: string;
    finalTotal: string;
    type: string;
    referenceNo: string;
    orderItemList: any[];
    print:boolean=false;
}
 